ref(node): Improve Node manual test logging (#5088)
This makes some small changes to the logs emitted by the manual tests in the node package, to make it a little cleaner and easier to see what's what.

Included changes:
- Use `colorize` function from release health tests in all test suites.
- Capture and display console logs in express test.
- Standardize success message across test suites.
- Use whitespace to separate each suite's logs from the next.
