+++ packages/node/test/manual/release-health/runner.js
@@ -1,21 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const { spawn } = require('child_process');
-
-const COLOR_RESET = '\x1b[0m';
-const COLORS = {
-  green: '\x1b[32m',
-  red: '\x1b[31m',
-  yellow: '\x1b[33m',
-};
-
-const colorize = (str, color) => {
-  if (!(color in COLORS)) {
-    throw new Error(`Unknown color. Available colors: ${Object.keys(COLORS).join(', ')}`);
-  }
-
-  return `${COLORS[color]}${str}${COLOR_RESET}`;
-};
+const { colorize } = require('../colorize');
 
 const scenariosDirs = ['session-aggregates', 'single-session'];
 const scenarios = [];
