+++ packages/node/test/manual/webpack-domain/index.js
@@ -1,4 +1,5 @@
 const Sentry = require('../../../build/cjs');
+const { colorize } = require('../colorize');
 
 let remaining = 2;
 
@@ -7,7 +8,7 @@ function makeDummyTransport() {
     --remaining;
 
     if (!remaining) {
-      console.error('SUCCESS: Webpack Node Domain test OK!');
+      console.log(colorize('PASSED: Webpack Node Domain test OK!\n', 'green'));
       process.exit(0);
     }
 
@@ -23,13 +24,13 @@ Sentry.init({
   beforeSend(event) {
     if (event.message === 'inside') {
       if (event.tags.a !== 'x' && event.tags.b !== 'c') {
-        console.error('FAILED: Scope contains incorrect tags');
+        console.log(colorize('FAILED: Scope contains incorrect tags\n', 'red'));
         process.exit(1);
       }
     }
     if (event.message === 'outside') {
       if (event.tags.a !== 'b') {
-        console.error('FAILED: Scope contains incorrect tags');
+        console.log(colorize('FAILED: Scope contains incorrect tags\n', 'red'));
         process.exit(1);
       }
     }
